.txt files for analysis
These files will be used in the analysis of manifestos.

These are the text files converted from PDF/Doc. Most are in Unicode utf-8. 
Special Note: JI_2013 was made through OCR (Optical Character Recognition) using the website: https://www.pdf2go.com/pdf-to-text . This is because the JI source file was made of images instead of text.